// src/app/shared/interface.ts
export interface App {
    // define state here
    coffeeList: Coffee[];
    cart: { name: string; quantity: number }[];
}

// export interface AppApp {
//     readonly app: App;
// }

export interface Coffee {
    name: string;
    price: number;
    recipe: RecipeItem[];
}

export interface RecipeItem {
    name: string;
    quantity: number;
}


//****************** 

export interface Remix {
    // define state here
    runningNo: number;
}

// export interface RemixRemix {
//     readonly remix: Remix;
// }